pyproject.toml
# Build project configuration for release
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "weco"
authors = [
    {name = "WeCo AI Team", email = "dhruv@weco.ai"},
]
description = "A client facing API for interacting with the WeCo AI function builder service."
readme = "README.md"
version = "0.1.8"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = ["asyncio", "httpx[http2]", "pillow"]
keywords = ["AI", "LLM", "machine learning", "data science", "function builder", "AI function"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License"
]

[project.urls]
Homepage = "https://github.com/WecoAI/weco-python"

[project.optional-dependencies]
dev = ["flake8", "flake8-pyproject", "black", "isort", "pytest-asyncio", "pytest-xdist", "build", "setuptools_scm"]

[tool.setuptools]
packages = ["weco"]

[tool.setuptools_scm]

# Linting code
[tool.flake8]
count = true
select = ["E9", "F63", "F7", "F82"]
show-source = true
statistics = true
max-line-length = 127
max-complexity = 10
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.env
  | build
  | dist
)/
'''

# Formatting code
[tool.black]
line-length = 127
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.env
  | build
  | dist
)/
'''

# Sorting imports
[tool.isort]
profile = "black"
line_length = 127
